Move post-asset-completion code to individual functions with a lookup table
